+++ apps/client/src/features/app-settings/panel/integrations-panel/integrationUtils.ts
@@ -1,0 +1,8 @@
+export const cycles = [
+  { id: 1, label: 'On Load', value: 'onLoad' },
+  { id: 2, label: 'On Start', value: 'onStart' },
+  { id: 3, label: 'On Pause', value: 'onPause' },
+  { id: 4, label: 'On Stop', value: 'onStop' },
+  { id: 5, label: 'Every second', value: 'onUpdate' },
+  { id: 6, label: 'On Finish', value: 'onFinish' },
+];
